Add FrameIdRange to FrameKinematicsVector to access all frame ids in the vector.
