+++ geometry/BUILD.bazel
@@ -67,6 +67,7 @@ drake_cc_library(
     ],
     deps = [
         ":geometry_ids",
+        ":utilities",
         "//common:autodiff",
         "//common:essential",
         "//common:symbolic",
@@ -172,6 +173,7 @@ drake_cc_library(
         ":internal_frame",
         ":internal_geometry",
         ":proximity_engine",
+        ":utilities",
     ],
 )
