+++ geometry/geometry_state.h
@@ -17,6 +17,7 @@
 #include "drake/geometry/internal_frame.h"
 #include "drake/geometry/internal_geometry.h"
 #include "drake/geometry/proximity_engine.h"
+#include "drake/geometry/utilities.h"
 
 namespace drake {
 namespace geometry {
@@ -25,44 +26,6 @@ namespace geometry {
 namespace internal {
 
 class GeometryVisualizationImpl;
-
-// A const range iterator through the keys of an unordered map.
-template <typename K, typename V>
-class MapKeyRange {
- public:
-  DRAKE_DEFAULT_COPY_AND_MOVE_AND_ASSIGN(MapKeyRange)
-
-  class ConstIterator {
-   public:
-    DRAKE_DEFAULT_COPY_AND_MOVE_AND_ASSIGN(ConstIterator)
-
-    const K& operator*() const { return itr_->first; }
-    const ConstIterator& operator++() {
-      ++itr_;
-      return *this;
-    }
-    bool operator!=(const ConstIterator& other) { return itr_ != other.itr_; }
-
-   private:
-    explicit ConstIterator(
-        typename std::unordered_map<K, V>::const_iterator itr)
-        : itr_(itr) {}
-
-   private:
-    typename std::unordered_map<K, V>::const_iterator itr_;
-    friend class MapKeyRange;
-  };
-
-  explicit MapKeyRange(const std::unordered_map<K, V>* map)
-      : map_(map) {
-    DRAKE_DEMAND(map);
-  }
-  ConstIterator begin() const { return ConstIterator(map_->cbegin()); }
-  ConstIterator end() const { return ConstIterator(map_->cend()); }
-
- private:
-  const std::unordered_map<K, V>* map_;
-};
 
 }  // namespace internal
 #endif
